# Optimize quantized max pool 2d